+++ demo/apps/basic/ui.r
@@ -1,5 +1,7 @@
-shinyUI(pageWithSidebar(
-  div(),
+library(ggvis)
+
+
+fluidPage(sidebarLayout(
   sidebarPanel(
     sliderInput("n", "Number of points", min = 1, max = nrow(mtcars),
                 value = 10, step = 1),